+++ infrastructure/aws/cdk/app.py
@@ -1,7 +1,7 @@
 """Construct App."""
 
 import os
-from typing import Any, Dict, List, Optional
+from typing import Any, List, Optional
 
 from aws_cdk import App, CfnOutput, Duration, Stack, Tags, aws_lambda
 from aws_cdk import aws_apigatewayv2 as apigw
@@ -42,7 +42,6 @@ class LambdaStack(Stack):
         runtime: aws_lambda.Runtime = aws_lambda.Runtime.PYTHON_3_10,
         concurrent: Optional[int] = None,
         permissions: Optional[List[iam.PolicyStatement]] = None,
-        environment: Optional[Dict] = None,
         role_arn: Optional[str] = None,
         context_dir: str = "../../",
         **kwargs: Any,
@@ -51,7 +50,6 @@ class LambdaStack(Stack):
         super().__init__(scope, id, *kwargs)
 
         permissions = permissions or []
-        environment = environment or {}
 
         iam_reader_role = None
         if role_arn:
@@ -60,6 +58,15 @@ class LambdaStack(Stack):
                 "veda-reader-dev-role",
                 role_arn=role_arn,
             )
+
+        lambda_env = {
+            **DEFAULT_ENV,
+            "TITILER_CMR_ROOT_PATH": app_settings.root_path,
+            "TITILER_CMR_S3_AUTH_STRATEGY": app_settings.s3_auth_strategy,
+        }
+
+        if app_settings.aws_request_payer:
+            lambda_env["AWS_REQUEST_PAYER"] = app_settings.aws_request_payer
 
         lambda_function = aws_lambda.Function(
             self,
@@ -74,11 +81,7 @@ class LambdaStack(Stack):
             memory_size=memory,
             reserved_concurrent_executions=concurrent,
             timeout=Duration.seconds(timeout),
-            environment={
-                **DEFAULT_ENV,
-                **environment,
-                "TITILER_CMR_ROOT_PATH": app_settings.root_path,
-            },
+            environment=lambda_env,
             log_retention=logs.RetentionDays.ONE_WEEK,
             role=iam_reader_role,
         )
@@ -154,7 +157,6 @@ lambda_stack = LambdaStack(
     concurrent=app_settings.max_concurrent,
     role_arn=app_settings.role_arn,
     permissions=perms,
-    environment=stack_settings.additional_env,
 )
 # Tag infrastructure
 for key, value in {